feat: simplify the collection page components
